TST: split test_accessor.py to a package
